Replace unsigned with uint16_t. Improve code documentation.

+++ src/lib/operators/table_scan.cpp
@@ -38,8 +38,9 @@ std::string &TableScan::replace_all(std::string &str, const std::string &old_val
 }
 
 std::shared_ptr<AbstractOperator> TableScan::recreate(const std::vector<AllParameterVariant> &args) const {
+  // Replace value in the new operator, if it's a parameter and an argument is available.
   if (_value.type() == typeid(ParameterValue)) {
-    unsigned index = boost::get<ParameterValue>(_value).index();
+    uint16_t index = boost::get<ParameterValue>(_value).index();
     if (index < args.size()) {
       return std::make_shared<TableScan>(_input_left->recreate(args), _column_name, _op, args[index], _value2);
     }
